Api/JewerlyStore/Controllers/UserController.cs
using JewelryStore.Services.Contracts;
using JewelryStore.Services.Dtos;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

namespace JewelryStore.Controllers
{
    public class UserController : BaseController
    {
        private readonly IUserService _userService;

        public UserController(IUserService userService)
        {
            _userService = userService;
        }

        [AllowAnonymous]
        [HttpPost]
        [Route("authenticate")]
        public IActionResult Authenticate([FromBody] AuthRequestDto authRequest)
        {
            return Ok(_userService.AuthenticateUser(authRequest));
        }
    }
}
